split app into lists and login
